% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCRPS_with_INLA.R
\name{score.inla}
\alias{score.inla}
\title{scores for inla}
\usage{
\method{score}{inla}(
  inla_result,
  scoring_rule = "scrps",
  n = 10000,
  test_data = NULL,
  full_y = NULL,
  test_idx = NULL,
  only_test_data = TRUE,
  parallelize = FALSE,
  n.cores = parallel::detectCores() - 1
)
}
\arguments{
\item{inla_result}{b}

\item{n}{b}

\item{test_data}{b}

\item{full_y}{b}

\item{test_idx}{b}

\item{only_test_data}{b}

\item{parallelize}{b}

\item{n.cores}{b}
}
\value{
b
}
\description{
scores for inla
}
\examples{
#to be completed
}
