% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCRPS_with_INLA.R
\name{score}
\alias{score}
\title{Generic score function}
\usage{
score(x, ...)
}
\arguments{
\item{x}{b}

\item{...}{b}
}
\value{
b
}
\description{
Generic score function
}
\examples{
#to be completed
}
