% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCRPS_with_INLA.R
\name{scrps.inlaa}
\alias{scrps.inlaa}
\alias{scrps__inlaa}
\title{SCRPS for INLA objects}
\usage{
scrps__inlaa(
  inla_result,
  n = 10000,
  test_data = NULL,
  full_y = NULL,
  test_idx = NULL,
  only_test_data = TRUE,
  parallelize = FALSE,
  n.cores = parallel::detectCores() - 1
)
}
\arguments{
\item{inla_result}{inla result object}

\item{n}{bla}

\item{test_data}{ble}

\item{full_y}{bli}

\item{test_idx}{blo}

\item{parallelize}{blu}

\item{n.cores}{bla}
}
\value{
SCRPS
}
\description{
SCRPS for INLA objects
}
\examples{

print("To be completed")
}
