% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCRPS.R
\name{SCRPS_2pnorm}
\alias{SCRPS_2pnorm}
\title{SCRPS for two-piece normal distribution}
\usage{
SCRPS_2pnorm(y, mu = 0, sd1 = 1, sd2 = 1)
}
\arguments{
\item{y}{Vector of observations.}

\item{mu}{Vector of location parameters.}

\item{sd1}{Vector of left-hand-side scale parameters. Must be positive.}

\item{sd2}{Vector of right-hand-side scale parameters. Must be positive.}
}
\value{
Vector of scores.
}
\description{
Calculates the SCRPS for observations \code{y} and two-piece normal distributions with location parameters = \code{mu} and left-hand-side scale parameters = \code{sd1} and right-hand-side scale parameters = \code{sd2}.
}
\examples{
SCRPS_2pnorm(7, 6, 2, 3)
}
