% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCRPS.R
\name{SCRPS_beta}
\alias{SCRPS_beta}
\title{SCRPS for beta distribution}
\usage{
SCRPS_beta(y, shape1, shape2)
}
\arguments{
\item{y}{Vector of observations.}

\item{shape1}{Vector of shape1 parameters. Must be positive.}

\item{shape2}{Vector of shape2 parameters. Must be positive.}
}
\value{
Vector of scores.
}
\description{
Calculates the SCRPS for observations \code{y} and beta distribution with shape1 parameter \code{shape1} and shape2 parameter \code{shape2}.
}
\examples{
SCRPS_beta(1,1,1)
}
