% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCRPS.R
\name{SCRPS_exp}
\alias{SCRPS_exp}
\title{SCRPS for the exponential distribution}
\usage{
SCRPS_exp(y, rate = 1)
}
\arguments{
\item{y}{Vector of observations.}

\item{rate}{Vector of rate parameters. Must be positive.}
}
\value{
Vector of scores.
}
\description{
Calculates the SCRPS for observations \code{y} and exponential distributions with rates = \code{rate}.
}
\examples{
SCRPS_exp(2, rate = 3)
}
