% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCRPS.R
\name{SCRPS_exp2}
\alias{SCRPS_exp2}
\title{SCRPS for exponential distribution with flexible location and scale}
\usage{
SCRPS_exp2(y, rate = 1, location = 0, scale = 1)
}
\arguments{
\item{y}{Vector of observations.}

\item{rate}{Vector of rate parameters. Must be positive.}

\item{location}{Vector of location parameters.}

\item{scale}{Vector of scale parameters. Must be positive.}
}
\value{
Vector of scores.
}
\description{
Calculates the SCRPS for observations \code{y} and exponential distributions with rate parameters = \code{rate}, location parameters = \code{location} and scale parameters = \code{scale}.
}
\examples{
SCRPS_exp2(10, rate = 2, location = 5, scale = 3)
}
