% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCRPS.R
\name{SCRPS_lapl}
\alias{SCRPS_lapl}
\title{SCRPS for the laplace distribution}
\usage{
SCRPS_lapl(y, mu = 0, b = 1)
}
\arguments{
\item{y}{Vector of observations.}

\item{mu}{Vector of location parameters.}

\item{b}{Vector of scale parameters. Must be positive.}
}
\value{
Vector of scores.
}
\description{
Calculates the SCRPS for observations \code{y} and laplace distributions with location parameters = \code{mu} and scale parameters = \code{b}.
}
\examples{
SCRPS_lapl(2, mu = 3, b = 2)
}
