% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCRPS.R
\name{SCRPS_lnorm}
\alias{SCRPS_lnorm}
\title{SCRPS for log-normal distribution}
\usage{
SCRPS_lnorm(y, meanlog = 0, sdlog = 1)
}
\arguments{
\item{y}{Vector of observations.}

\item{meanlog}{Vector of locationlog parameters.}

\item{sdlog}{Vector of scalelog parameters. Must be positive.}
}
\value{
Vector of scores.
}
\description{
Calculates the SCRPS for observations \code{y} and lognormal distribution with locationlog parameter \code{meanlog} and scalelog parameter \code{sdlog}.
}
\examples{
SCRPS_lnorm(1,2,3)
}
