% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCRPS.R
\name{SCRPS_logis}
\alias{SCRPS_logis}
\title{SCRPS for logistic distribution}
\usage{
SCRPS_logis(y, location = 0, scale = 1)
}
\arguments{
\item{y}{Vector of observations.}

\item{location}{Vector of location parameters.}

\item{scale}{Vector of scale parameters. Must be positive.}
}
\value{
Vector of scores.
}
\description{
Calculates the SCRPS for observations \code{y} and logistic distributions with location parameters = \code{location} and scale parameters = \code{scale}.
}
\examples{
SCRPS_logis(3,location = -5, scale = 2)
}
