% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCRPS.R
\name{SCRPS_t}
\alias{SCRPS_t}
\title{SCRPS for Student's t distribution}
\usage{
SCRPS_t(y, df, location = 0, scale = 1)
}
\arguments{
\item{y}{Vector of observations.}

\item{df}{Vector of degrees of freedom parameters. Must be greater than 1.}

\item{location}{Vector of location parameters.}

\item{scale}{Vector of scale parameters. Must be positive.}
}
\value{
Vector of scores.
}
\description{
Calculates the SCRPS for observations \code{y} and Student's t distribution with degrees of freedom parameter \code{df} and optionally location parameter \code{location} and scale parameter \code{scale}.
}
\examples{
SCRPS_t(1,2,3,4)

}
