% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCRPS.R
\name{SCRPS_unif}
\alias{SCRPS_unif}
\title{SCRPS for the uniform distribution}
\usage{
SCRPS_unif(y, min = 0, max = 1)
}
\arguments{
\item{y}{Vector of observations.}

\item{min}{Vector of lower limits.}

\item{max}{Vector of upper limits.}
}
\value{
Vector of scores.
}
\description{
Calculates the SCRPS for observations \code{y} and uniform distributions with lower limits = \code{min} and upper limits = \code{max}.
}
\examples{
SCRPS_unif(3, min = 10, max = 17)
}
