% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCRPS.R
\name{SCRPS_gamma}
\alias{SCRPS_gamma}
\title{SCRPS for Gamma distribution}
\usage{
SCRPS_gamma(y, shape = 1, rate = 1, scale = 1/rate)
}
\arguments{
\item{y}{bla}

\item{shape}{bla}

\item{rate}{bla}

\item{scale}{bla}
}
\value{
bla
}
\description{
Calculates the SCRPS for observations \code{y} and
}
\examples{
#' SCRPS_gamma(7, 6, 2, 3)
}
